With edit

+++ ShoppingListExam/app/src/main/java/com/example/ShoppingListExam/Adapters/ProductAdapter.kt
@@ -2,6 +2,8 @@ package com.example.ShoppingListExam.Adapters
 
 
 
+import android.content.Context
+import android.content.Intent
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -9,13 +11,16 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.ShoppingListExam.EditActivity
 import com.example.ShoppingListExam.R
+import com.example.ShoppingListExam.SettingsActivity
+import com.example.ShoppingListExam.ShareActivity
 import com.example.ShoppingListExam.data.DAOProducts
 import com.example.ShoppingListExam.data.Product
+import com.example.ShoppingListExam.data.Repository
 
 
-
-class ProductAdapter(var products: MutableList<Product>) :
+class ProductAdapter(var products: MutableList<Product>, var context: Context) :
     RecyclerView.Adapter<ProductAdapter.ViewHolder>() {
 
 
@@ -24,7 +29,7 @@ class ProductAdapter(var products: MutableList<Product>) :
         val inf = LayoutInflater.from(parent.context).inflate(R.layout.card_layout, parent, false)
 
 
-        return ViewHolder(inf)
+        return ViewHolder(inf,context)
 
     }
 
@@ -38,11 +43,12 @@ class ProductAdapter(var products: MutableList<Product>) :
         return products.size
     }
 
-    inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView),
+    inner class ViewHolder(itemView: View,context: Context) : RecyclerView.ViewHolder(itemView),
         View.OnClickListener {
 
         val productQuantity: TextView = itemView.findViewById(R.id.product_quantity)
         val productOptions: ImageView = itemView.findViewById(R.id.product_options)
+        val productEdit:ImageView = itemView.findViewById(R.id.product_edit)
         val productTitle: TextView = itemView.findViewById(R.id.product_title)
         val productDetail: TextView = itemView.findViewById(R.id.product_detail)
         fun bind(product: Product) {
@@ -56,17 +62,31 @@ class ProductAdapter(var products: MutableList<Product>) :
 
         init {
             productOptions.setOnClickListener(this)
+            productEdit.setOnClickListener(this)
+
 
         }
+
+
 
         override fun onClick(view: View) {
 
             val position = adapterPosition
-
-            if (position != RecyclerView.NO_POSITION) {
+            val dao = DAOProducts()
+            if (position != RecyclerView.NO_POSITION && view == productOptions ) {
                 Log.d("click", position.toString())
-                val dao = DAOProducts()
                 dao.delete(products[position].key)
+            }
+            if (position != RecyclerView.NO_POSITION && view == productEdit ) {
+                Log.d("click", position.toString())
+                val key = products[position].key
+                ///sharing
+               // val editIntent = Intent(this, ShareActivity::class.java)
+                val editIntent = context.startActivity(Intent(context, EditActivity::class.java).putExtra("key",key))
+
+                Log.d("stringlist",key)
+                /// run intent with edit
+                Log.d("click","editlii")
             }
 
         }
